+++ src/m_config.c
@@ -181,7 +181,7 @@ static default_t cvars[] =
     CONFIG_VARIABLE_INT          (r_textures,                       r_textures,                            r_textures,                            BOOLVALUEALIAS        ),
     CONFIG_VARIABLE_INT          (r_translucency,                   r_translucency,                        r_translucency,                        BOOLVALUEALIAS        ),
     CONFIG_VARIABLE_INT          (s_channels,                       s_channels,                            s_channels,                            NOVALUEALIAS          ),
-    CONFIG_VARIABLE_INT          (s_menumusicvolume,                s_menumusicvolume,                     s_menumusicvolume,                     BOOLVALUEALIAS        ),
+    CONFIG_VARIABLE_INT          (s_lowermenumusic,                 s_lowermenumusic,                      s_lowermenumusic,                      BOOLVALUEALIAS        ),
     CONFIG_VARIABLE_INT_PERCENT  (s_musicvolume,                    s_musicvolume,                         s_musicvolume,                         NOVALUEALIAS          ),
     CONFIG_VARIABLE_INT          (s_randommusic,                    s_randommusic,                         s_randommusic,                         BOOLVALUEALIAS        ),
     CONFIG_VARIABLE_INT          (s_randompitch,                    s_randompitch,                         s_randompitch,                         BOOLVALUEALIAS        ),
@@ -924,8 +924,8 @@ static void M_CheckCVARs(dboolean ispackageconfig)
 
     s_channels = BETWEEN(s_channels_min, s_channels, s_channels_max);
 
-    if (s_menumusicvolume != false && s_menumusicvolume != true)
-        s_menumusicvolume = s_menumusicvolume_default;
+    if (s_lowermenumusic != false && s_lowermenumusic != true)
+        s_lowermenumusic = s_lowermenumusic_default;
 
     s_musicvolume = BETWEEN(s_musicvolume_min, s_musicvolume, s_musicvolume_max);
     musicVolume = (s_musicvolume * 31 + 50) / 100;