+++ src/m_config.h
@@ -160,7 +160,7 @@ extern dboolean     r_supersampling;
 extern dboolean     r_textures;
 extern dboolean     r_translucency;
 extern int          s_channels;
-extern dboolean     s_menumusicvolume;
+extern dboolean     s_lowermenumusic;
 extern int          s_musicvolume;
 extern dboolean     s_randommusic;
 extern dboolean     s_randompitch;
@@ -655,7 +655,7 @@ enum
 #define s_channels_default                      32
 #define s_channels_max                          64
 
-#define s_menumusicvolume_default               true
+#define s_lowermenumusic_default                true
 
 #define s_musicvolume_min                       0
 #define s_musicvolume_default                   100